Moved pre processing data steps from the dashboard to a post processing step in the webscraper
