+++ exePostProcessData.bat
@@ -1,0 +1,1 @@
+call python webscraper\prg_postprocess_data.py
